refactor(agents): the boot-time door splits from the request path
registry.go crossed the 500-line cap. The split is along the boundary that
already exists rather than a convenient one: registration.go is everything
that runs at BOOT against a spec, before any request exists, and registry.go
is what happens to a call.
